Extract slash action executor
