+++ Sources/QuillCodeApp/WorkspaceModel.swift
@@ -1497,72 +1497,7 @@ public final class QuillCodeWorkspaceModel {
         refreshTopBar(agentStatus: TopBarAgentStatusLabel.idle)
     }
 
-    private func runSlashCommandDispatchAction(_ action: WorkspaceSlashCommandDispatchAction, workspaceRoot: URL) {
-        switch action {
-        case .appendTranscript(let transcript):
-            appendLocalCommandTranscript(transcript)
-        case .newChat:
-            _ = newChat()
-        case .setMode(let mode, let userText):
-            setMode(mode)
-            appendLocalCommandTranscript(WorkspaceSlashCommandTranscriptPlanner.mode(
-                userText: userText,
-                mode: mode
-            ))
-        case .setModel(let model, let userText):
-            let modelID = setModel(model)
-            appendLocalCommandTranscript(WorkspaceSlashCommandTranscriptPlanner.model(
-                userText: userText,
-                model: modelID
-            ))
-        case .renameThread(let title, let userText):
-            let succeeded = root.selectedThreadID.map { renameThread($0, to: title) } ?? false
-            appendLocalCommandTranscript(WorkspaceSlashCommandTranscriptPlanner.renameThread(
-                userText: userText,
-                requestedTitle: title,
-                succeeded: succeeded
-            ))
-        case .renameProject(let name, let userText):
-            let succeeded = root.selectedProjectID.map { renameProject($0, to: name) } ?? false
-            appendLocalCommandTranscript(WorkspaceSlashCommandTranscriptPlanner.renameProject(
-                userText: userText,
-                requestedName: name,
-                succeeded: succeeded
-            ))
-        case .addSSHProject(let address, let userText):
-            if let projectID = addSSHProject(address),
-               let project = root.projects.first(where: { $0.id == projectID }) {
-                appendLocalCommandTranscript(WorkspaceSlashCommandTranscriptPlanner.sshProjectAdded(
-                    userText: userText,
-                    projectName: project.name,
-                    displayPath: project.displayPath
-                ))
-            } else {
-                appendLocalCommandTranscript(WorkspaceSlashCommandTranscriptPlanner.sshProjectFailed(
-                    userText: userText,
-                    message: lastError
-                ))
-            }
-        case .remember(let content, let userText):
-            runRememberSlashCommand(content, originalPrompt: userText)
-        case .threadFollowUp(let scheduleText, let userText):
-            runThreadFollowUpSlashCommand(scheduleText, originalPrompt: userText)
-        case .workspaceSchedule(let scheduleText, let userText):
-            runWorkspaceScheduleSlashCommand(scheduleText, originalPrompt: userText)
-        case .workspaceCommand(let commandID, let userText):
-            if !runWorkspaceCommand(commandID, workspaceRoot: workspaceRoot) {
-                appendLocalCommandTranscript(WorkspaceSlashCommandTranscriptPlanner.workspaceCommandFailed(
-                    userText: userText
-                ))
-            }
-        case .toolCall(let call):
-            _ = runToolCall(call, workspaceRoot: workspaceRoot)
-        case .environmentAction(let query, let userText):
-            runEnvironmentSlashCommand(query, originalPrompt: userText, workspaceRoot: workspaceRoot)
-        }
-    }
-
-    private func runThreadFollowUpSlashCommand(_ scheduleText: String, originalPrompt: String) {
+    func runThreadFollowUpSlashCommand(_ scheduleText: String, originalPrompt: String) {
         appendScheduledAutomationTranscript(
             createThreadFollowUpAutomation(matching: scheduleText),
             success: {
@@ -1580,7 +1515,7 @@ public final class QuillCodeWorkspaceModel {
         )
     }
 
-    private func runWorkspaceScheduleSlashCommand(_ scheduleText: String, originalPrompt: String) {
+    func runWorkspaceScheduleSlashCommand(_ scheduleText: String, originalPrompt: String) {
         appendScheduledAutomationTranscript(
             createWorkspaceScheduleAutomation(matching: scheduleText),
             success: {
@@ -1609,7 +1544,7 @@ public final class QuillCodeWorkspaceModel {
         appendLocalCommandTranscript(transcript)
     }
 
-    private func runRememberSlashCommand(_ content: String, originalPrompt: String) {
+    func runRememberSlashCommand(_ content: String, originalPrompt: String) {
         let mutation = WorkspaceMemoryEngine.saveGlobal(
             content: content,
             userText: originalPrompt,
@@ -1618,7 +1553,7 @@ public final class QuillCodeWorkspaceModel {
         applyGlobalMemoryMutation(mutation)
     }
 
-    private func runEnvironmentSlashCommand(_ query: String?, originalPrompt: String, workspaceRoot: URL) {
+    func runEnvironmentSlashCommand(_ query: String?, originalPrompt: String, workspaceRoot: URL) {
         refreshProjectMetadata(root.selectedProjectID)
         let plan = WorkspaceEnvironmentSlashCommandPlanner.plan(
             query: query,
@@ -1633,7 +1568,7 @@ public final class QuillCodeWorkspaceModel {
         }
     }
 
-    private func appendLocalCommandTranscript(_ transcript: WorkspaceLocalCommandTranscript) {
+    func appendLocalCommandTranscript(_ transcript: WorkspaceLocalCommandTranscript) {
         if selectedThread == nil {
             _ = newChat()
         }
