+++ docs/DECISIONS.md
@@ -237,6 +237,7 @@
 - Runtime/auth status labels live in `QuillCodeRuntimeStatusLabel`, separate from lifecycle labels. `RuntimeFactory`, `WorkspaceRuntimeIssueBuilder`, and the desktop sign-in controller should use those constants for mock, sign-in-needed, developer-key-needed, signed-in, ready, and sign-in-failed statuses so runtime issue detection does not depend on scattered string sentinels.
 - Agent progress status copy lives in `WorkspaceAgentStatusBuilder`. `WorkspaceModel` owns when agent progress arrives and when the top bar refreshes, while the builder owns event-kind-to-status labels such as Queued, Running, Review, Streaming, Finishing, and Failed so progress copy stays directly tested and aligned with top-bar presentation semantics.
 - Slash-command local transcript copy lives in `WorkspaceSlashCommandTranscriptPlanner`. `WorkspaceModel` still owns command side effects such as renaming, scheduling, SSH project creation, memory writes, local environment action execution, and tool dispatch, but the user-visible success/failure records for `/help`, `/status`, `/mode`, `/model`, `/rename`, `/project rename`, `/ssh`, `/follow-up`, `/workspace-check`, `/env`, invalid commands, and unknown commands are pure planner outputs with focused tests.
+- Slash-command dispatch is split into parser, planner, and executor stages. `SlashCommandParser` and its family parsers turn text into parsed commands, `WorkspaceSlashCommandDispatchPlanner` maps parsed commands to typed `WorkspaceSlashCommandDispatchAction` values, and `WorkspaceSlashCommandActionExecutor` applies those typed actions through the workspace model. `WorkspaceModel.handleSlashCommand` should stay as lifecycle coordination only: plan, execute, clear send state, and refresh the top bar.
 - Pull request slash-command parsing lives in `SlashPullRequestCommandParser`. `SlashCommandParser` owns only the top-level `/pr` delegation, while PR subcommands, selector/body splitting, reviewer and label argument construction, and merge flags stay beside focused parser tests.
 - Thread lifecycle slash-command parsing lives in `SlashThreadCommandParser`. `SlashCommandParser` owns only top-level thread alias delegation, while `/new`, `/compact`, `/rename`, `/duplicate`, `/archive`, `/unarchive` aliases, command IDs, and thread rename usage copy stay beside focused parser tests.
 - Memory slash-command parsing lives in `SlashMemoryCommandParser`. `SlashCommandParser` owns only top-level memory alias delegation, while `/memory`, `/memories`, `/remember`, memory pane command IDs, and remember-content trimming stay beside focused parser tests.
